Add filter in one Rest method with specification in service

+++ src/main/java/smida/test_task/havriushenko/stockRegestry/models/StockModel.java
@@ -22,7 +22,7 @@ public class StockModel {
     @Column(name = "usreou")
     private long USREOU;            //Unified State Register of Enterprises and Organizations of Ukraine
     @Column
-    private double quantity;
+    private long quantity;
     @Column(name = "total_nominal_value")
     private double totalNominalValue;
     @Column(name = "nominal_value")
